+++ PERUSTARS/PERUSTARS/ProfileManagement/Domain/Model/Events/ProfileRegisteredEvent.cs
@@ -1,12 +1,12 @@
-using System.Numerics;
+
 using MediatR;
 
 namespace PERUSTARS.ProfileManagement.Domain.Model.Events
 {
     public class ProfileRegisteredEvent: INotification
     {
-        public BigInteger ArtistId { get; set; }
-        public BigInteger HobbyistId { get; set; }
+        public long ArtistId { get; set; }
+        public long HobbyistId { get; set; }
         
     }
 }
